pyproject.toml
[project]
name = "clean-architecture-fastapi"
version = "0.1.0"
description = "FastAPI application following Clean Architecture principles"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.17.2",
    "fastapi[standard]>=0.121.2",
    "pwdlib[argon2]>=0.3.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.44",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
]
